Refactorizacion en carpetas de clean arquitecture
