+++ back-end/back-end/Application/EMPTY.txt
@@ -1,0 +1,1 @@
+Application business rules
