+++ back-end/back-end/Domain/CompanyModel.cs
