+++ back-end/back-end/Domain/LoginUserModel.cs
@@ -1,0 +1,13 @@
+﻿namespace back_end.Models
+{
+    public class LoginUserModel
+    {
+        public LoginUserModel()
+        {
+            this.NicknameOrEmail = "";
+            this.Password = "";
+        }
+        public string NicknameOrEmail { get; set; }
+        public string Password { get; set; }
+    }
+}
